move parse_requirements_file to pants.util
