+++ src/python/pants/backend/python/goals/lockfile.py
@@ -9,7 +9,6 @@ from collections import defaultdict
 from dataclasses import dataclass
 from operator import itemgetter
 
-from pants.backend.python.pip_requirement import PipRequirement
 from pants.backend.python.subsystems.setup import PythonSetup
 from pants.backend.python.target_types import PythonRequirementResolveField, PythonRequirementsField
 from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
@@ -42,6 +41,7 @@ from pants.engine.unions import UnionRule
 from pants.util.docutil import bin_name
 from pants.util.logging import LogLevel
 from pants.util.ordered_set import FrozenOrderedSet
+from pants.util.pip_requirement import PipRequirement
 
 
 @dataclass(frozen=True)
