+++ src/python/pants/backend/python/macros/pipenv_requirements.py
@@ -11,7 +11,6 @@ from pants.backend.python.macros.common_fields import (
     TypeStubsModuleMappingField,
 )
 from pants.backend.python.macros.common_requirements_rule import _generate_requirements
-from pants.backend.python.pip_requirement import PipRequirement
 from pants.backend.python.subsystems.setup import PythonSetup
 from pants.backend.python.target_types import PythonRequirementResolveField, PythonRequirementTarget
 from pants.engine.rules import collect_rules, rule
@@ -24,6 +23,7 @@ from pants.engine.target import (
 )
 from pants.engine.unions import UnionMembership, UnionRule
 from pants.util.logging import LogLevel
+from pants.util.pip_requirement import PipRequirement
 
 
 class PipenvSourceField(SingleSourceField):
