+++ src/python/pants/backend/python/target_types.py
@@ -25,7 +25,6 @@ from typing import (
 from packaging.utils import canonicalize_name as canonicalize_project_name
 
 from pants.backend.python.macros.python_artifact import PythonArtifact
-from pants.backend.python.pip_requirement import PipRequirement
 from pants.backend.python.subsystems.setup import PythonSetup
 from pants.core.goals.generate_lockfiles import UnrecognizedResolveNamesError
 from pants.core.goals.package import OutputPathField
@@ -74,7 +73,7 @@ from pants.option.subsystem import Subsystem
 from pants.source.filespec import Filespec
 from pants.util.docutil import bin_name, doc_url, git_url
 from pants.util.frozendict import FrozenDict
-from pants.util.requirements import load_requirements
+from pants.util.pip_requirement import PipRequirement
 from pants.util.strutil import help_text, softwrap
 
 logger = logging.getLogger(__name__)
@@ -1314,12 +1313,6 @@ class PythonRequirementTarget(Target):
                     """
                 )
             )
-
-
-def parse_requirements_file(content: str, *, rel_path: str) -> Iterator[PipRequirement]:
-    """Parse all `PipRequirement` objects from a requirements.txt-style file."""
-    for line, i in load_requirements(content):
-        yield PipRequirement.parse(line, description_of_origin=f"{rel_path} at line {i}")
 
 
 # -----------------------------------------------------------------------------------------------
