+++ src/python/pants/backend/python/util_rules/pex_requirements_test.py
@@ -8,7 +8,6 @@ import textwrap
 
 import pytest
 
-from pants.backend.python.pip_requirement import PipRequirement
 from pants.backend.python.subsystems.setup import InvalidLockfileBehavior, PythonSetup
 from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
 from pants.backend.python.util_rules.lockfile_metadata import PythonLockfileMetadataV3
@@ -30,6 +29,7 @@ from pants.core.util_rules.lockfile_metadata import (
 from pants.engine.internals.native_engine import EMPTY_DIGEST
 from pants.testutil.option_util import create_subsystem
 from pants.util.ordered_set import FrozenOrderedSet
+from pants.util.pip_requirement import PipRequirement
 from pants.util.strutil import comma_separated_list
 
 METADATA = PythonLockfileMetadataV3(
