+++ src/python/pants/util/pip_requirement.py
